# courseGenie_backEnd/docs/apidocs/ui/openapi.json
{
  "openapi": "3.0.3",
  "info" : {
    "description" : "<h1>course-genie-backend<\/h1><p>Demo project for Spring Boot<\/p>",
    "version" : "0.0.1-SNAPSHOT",
    "title" : "course-genie-backend"
  },
  "servers" : [
    {
      "url" : "/"
    }
  ],
...
  "components" : {
    "schemas" : {
...
      "CLODTO" : {
"type" : "object",
"title" : "CLODTO",
"required" : ["courseId", "cloId"],
"properties" : {
  "courseId" : {
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  },
  "description" : {
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "name" : {
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "cloId" : {
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  }
},
"example" : {
  "courseId" : 12345,
  "description" : "...",
  "name" : "...",
  "cloId" : 12345
},
"description" : ""
      }
...
      "xml_ns0_CLODTO" : {
"type" : "object",
"title" : "CLODTO",
"required" : ["cloId", "courseId"],
"allOf" : [
  {
"properties" : {
  "cloId" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  },
  "courseId" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  },
  "description" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "name" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "string"
  }
}
  }
],
"description" : ""
      }
...
    }
  },
  "paths": {
...
    "\/api/assessments/{assassmentId}/clo/{cloId}" : {
      "post" : {
        "tags" : [ "AssessmentController" ],
        "operationId" : "mapCloToAssessment",
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
"description" : "",
"type" : "boolean"
                }
              }, 
              "application/xml" : {
                "schema" : {
"description" : "",
"type" : "boolean"
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
          {
            "name" : "assassmentId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "number"
            },
            "description" : ""
          },
          {
            "name" : "cloId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "number"
            },
            "description" : ""
          }
        ]
      },
      "put" : {
        "tags" : [ "AssessmentController" ],
        "operationId" : "unmapCloToAssessment",
        "responses" : {
          "204" : {
            "content" : {
              "application/json" : {
                "schema" : {
"description" : "",
"type" : "boolean"
                }
              }, 
              "application/xml" : {
                "schema" : {
"description" : "",
"type" : "boolean"
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
          {
            "name" : "assassmentId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "number"
            },
            "description" : ""
          },
          {
            "name" : "cloId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "number"
            },
            "description" : ""
          }
        ]
      }
    },
...
    "\/api/clos" : {
      "get" : {
        "tags" : [ "CLOController" ],
        "operationId" : "getAllClos",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "example" : {
                    "value" : "[ {\n  \"courseId\" : 12345,\n  \"description\" : \"...\",\n  \"name\" : \"...\",\n  \"cloId\" : 12345\n} ]"
                  }
                },
                "schema" : {
"description" : "",
"type" : "array",
"items" : {
  "$ref" : "#/components/schemas/CLODTO"
}
                }
              }, 
              "application/xml" : {
                "examples" : {
                  "example" : {
                    "value" : "<list>\n  <c-----\u003E\n    <cloId>...<\/cloId>\n    <description>...<\/description>\n    <name>...<\/name>\n    <courseId>...<\/courseId>\n  <\/c-----\u003E\n<\/list>\n"
                  }
                },
                "schema" : {
"description" : "",
"type" : "array",
"items" : {
  "$ref" : "#/components/schemas/xml_ns0_CLODTO"
}
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
        ]
      },
      "post" : {
        "tags" : [ "CLOController" ],
        "operationId" : "createClo",
        "requestBody" : {
          "required" : false,
          "content" : {
            "application/json" : {
              "examples" : {
                "example" : {
                  "value" : "{\n  \"courseId\" : 12345,\n  \"description\" : \"...\",\n  \"name\" : \"...\",\n  \"cloId\" : 12345\n}"
                }
              },
              "schema" : {
"$ref" : "#/components/schemas/CLODTO",
"description" : ""
              }
            }, 
            "application/xml" : {
              "examples" : {
                "example" : {
                  "value" : "<c-----\u003E\n  <cloId>...<\/cloId>\n  <description>...<\/description>\n  <name>...<\/name>\n  <courseId>...<\/courseId>\n<\/c-----\u003E\n"
                }
              },
              "schema" : {
"$ref" : "#/components/schemas/xml_ns0_CLODTO",
"description" : ""
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "example" : {
                    "value" : "{\n  \"courseId\" : 12345,\n  \"description\" : \"...\",\n  \"name\" : \"...\",\n  \"cloId\" : 12345\n}"
                  }
                },
                "schema" : {
"$ref" : "#/components/schemas/CLODTO",
"description" : ""
                }
              }, 
              "application/xml" : {
                "examples" : {
                  "example" : {
                    "value" : "<c-----\u003E\n  <cloId>...<\/cloId>\n  <description>...<\/description>\n  <name>...<\/name>\n  <courseId>...<\/courseId>\n<\/c-----\u003E\n"
                  }
                },
                "schema" : {
"$ref" : "#/components/schemas/xml_ns0_CLODTO",
"description" : ""
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
        ]
      },
      "put" : {
        "tags" : [ "CLOController" ],
        "operationId" : "updateClo",
        "requestBody" : {
          "required" : false,
          "content" : {
            "application/json" : {
              "examples" : {
                "example" : {
                  "value" : "{\n  \"courseId\" : 12345,\n  \"description\" : \"...\",\n  \"name\" : \"...\",\n  \"cloId\" : 12345\n}"
                }
              },
              "schema" : {
"$ref" : "#/components/schemas/CLODTO",
"description" : ""
              }
            }, 
            "application/xml" : {
              "examples" : {
                "example" : {
                  "value" : "<c-----\u003E\n  <cloId>...<\/cloId>\n  <description>...<\/description>\n  <name>...<\/name>\n  <courseId>...<\/courseId>\n<\/c-----\u003E\n"
                }
              },
              "schema" : {
"$ref" : "#/components/schemas/xml_ns0_CLODTO",
"description" : ""
              }
            }
          }
        },
        "responses" : {
          "204" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "example" : {
                    "value" : "{\n  \"courseId\" : 12345,\n  \"description\" : \"...\",\n  \"name\" : \"...\",\n  \"cloId\" : 12345\n}"
                  }
                },
                "schema" : {
"$ref" : "#/components/schemas/CLODTO",
"description" : ""
                }
              }, 
              "application/xml" : {
                "examples" : {
                  "example" : {
                    "value" : "<c-----\u003E\n  <cloId>...<\/cloId>\n  <description>...<\/description>\n  <name>...<\/name>\n  <courseId>...<\/courseId>\n<\/c-----\u003E\n"
                  }
                },
                "schema" : {
"$ref" : "#/components/schemas/xml_ns0_CLODTO",
"description" : ""
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
        ]
      }
    },
    "\/api/clos/{cloId}" : {
      "delete" : {
        "tags" : [ "CLOController" ],
        "operationId" : "deleteClo",
        "responses" : {
          "204" : {
            "content" : {
              "application/json" : {
                "schema" : {
"description" : "",
"type" : "boolean"
                }
              }, 
              "application/xml" : {
                "schema" : {
"description" : "",
"type" : "boolean"
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
          {
            "name" : "cloId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "number"
            },
            "description" : ""
          }
        ]
      },
      "get" : {
        "tags" : [ "CLOController" ],
        "operationId" : "getCloById",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "example" : {
                    "value" : "{\n  \"courseId\" : 12345,\n  \"description\" : \"...\",\n  \"name\" : \"...\",\n  \"cloId\" : 12345\n}"
                  }
                },
                "schema" : {
"$ref" : "#/components/schemas/CLODTO",
"description" : ""
                }
              }, 
              "application/xml" : {
                "examples" : {
                  "example" : {
                    "value" : "<c-----\u003E\n  <cloId>...<\/cloId>\n  <description>...<\/description>\n  <name>...<\/name>\n  <courseId>...<\/courseId>\n<\/c-----\u003E\n"
                  }
                },
                "schema" : {
"$ref" : "#/components/schemas/xml_ns0_CLODTO",
"description" : ""
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
          {
            "name" : "cloId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "number"
            },
            "description" : ""
          }
        ]
      }
    },
...
  }
}